+++ source/mysql/protocol/prepared.d
@@ -813,12 +813,9 @@ public:
 	If there are long data items among the expected result columns you can
 	specify that they are to be subject to chunked transfer via a delegate.
 
-	WARNING: This function is not currently unittested.
-
 	Params: csa = An optional array of ColumnSpecialization structs.
 	Returns: A (possibly empty) ResultSequence.
 	+/
-	//TODO: This needs unittested
 	ResultSequence querySequence(ColumnSpecialization[] csa = null)
 	{
 		enforceReadyForCommand();