+++ CHANGELOG.md
@@ -12,6 +12,8 @@ TBD - TBD
 	for prepared statements.
 - **New:**
 	Various new subclasses of MySQLException added, for better fine-grained control.
+- **New:**
+	Add `length` property and `opDollar` to struct Row.
 - **Change:**
 	Drop support for DMDFE 2.067.x and below. Compiles on
 	DMDFE 2.068.2 through 2.072.0. See [.travis.yml](https://github.com/mysql-d/mysql-native/blob/master/.travis.yml)